# app/Models/Category.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;
use Illuminate\Database\Eloquent\Relations\BelongsToMany;


/**
 * @OA\Schema(
 *     schema="Category",
 *     title="Category",
 *     description="Category model",
 *     @OA\Property(
 *         property="id",
 *         type="integer",
 *         description="ID de la catégorie"
 *     ),
 *     @OA\Property(
 *         property="title",
 *         type="string",
 *         description="Titre de la catégorie"
 *     ),
 *     @OA\Property(
 *         property="created_at",
 *         type="string",
 *         format="date-time",
 *         description="Date de création de la catégorie"
 *     ),
 *     @OA\Property(
 *         property="updated_at",
 *         type="string",
 *         format="date-time",
 *         description="Date de mise à jour de la catégorie"
 *     )
 * )
 */

class Category extends Model
{
    use HasFactory;

    protected $fillable =
        ["title",
    ];

    public function products(): BelongsToMany
    {
        return $this->belongsToMany(Product::class);
    }
}